// frontend/src/services/generated/models/ProgramSetGroupResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Fitlytics API
 * Workout tracking and training analytics API. Protected routes require a WorkOS access token in the `Authorization: Bearer ...` header.
 *
 * The version of the OpenAPI document: 0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ProgramSetResponse } from './ProgramSetResponse';
import {
    ProgramSetResponseFromJSON,
    ProgramSetResponseFromJSONTyped,
    ProgramSetResponseToJSON,
    ProgramSetResponseToJSONTyped,
} from './ProgramSetResponse';

/**
 * 
 * @export
 * @interface ProgramSetGroupResponse
 */
export interface ProgramSetGroupResponse {
    /**
     * 
     * @type {string}
     * @memberof ProgramSetGroupResponse
     */
    id?: string;
    /**
     * 
     * @type {number}
     * @memberof ProgramSetGroupResponse
     */
    sequence?: number;
    /**
     * 
     * @type {Array<ProgramSetResponse>}
     * @memberof ProgramSetGroupResponse
     */
    sets?: Array<ProgramSetResponse>;
}

/**
 * Check if a given object implements the ProgramSetGroupResponse interface.
 */
export function instanceOfProgramSetGroupResponse(value: object): value is ProgramSetGroupResponse {
    return true;
}

export function ProgramSetGroupResponseFromJSON(json: any): ProgramSetGroupResponse {
    return ProgramSetGroupResponseFromJSONTyped(json, false);
}

export function ProgramSetGroupResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProgramSetGroupResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'sequence': json['sequence'] == null ? undefined : json['sequence'],
        'sets': json['sets'] == null ? undefined : ((json['sets'] as Array<any>).map(ProgramSetResponseFromJSON)),
    };
}

export function ProgramSetGroupResponseToJSON(json: any): ProgramSetGroupResponse {
    return ProgramSetGroupResponseToJSONTyped(json, false);
}

export function ProgramSetGroupResponseToJSONTyped(value?: ProgramSetGroupResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'sequence': value['sequence'],
        'sets': value['sets'] == null ? undefined : ((value['sets'] as Array<any>).map(ProgramSetResponseToJSON)),
    };
}